MLOps Azure/requirements.txt
# Core ML libraries
scikit-learn==1.7.2 
pandas==2.0.0
numpy==1.24.0

# NLP libraries
nltk==3.8.1
spacy==3.7.0
contractions==0.1.73

# MLflow for experiment tracking
mlflow==2.9.0

# FastAPI for REST API
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-multipart==0.0.6

# Configuration
pyyaml==6.0

# Azure SDK
azure-ai-ml==1.11.0
azure-identity==1.15.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.0

# Testing
pytest==7.4.0
httpx==0.25.0
